Add files via upload
Users can log into the system by entering their email address or username and password on this user-friendly login page. There is a "Remember Me" checkbox on it that lets users save their login details for later sessions. The responsive and simple design makes sure it appears well on desktop and mobile platforms. To guarantee that users provide legitimate credentials before submitting the form, basic validation is in place. In order to provide functionality and user-friendliness, the page can additionally establish a connection with an API for authentication needs.